requirements.txt
numpy==1.26.*
pandas==2.1.*
scikit-learn==1.3.*
matplotlib==3.8.*
pyyaml==6.0.*
torch==2.1.*
tqdm==4.66.*
